+++ src/3rdparty/opennurbs/opennurbs_texture.h
@@ -126,7 +126,7 @@ public:
     default_channel = 1,	   // Use either default mapping, or the "Custom"
 							               // mapping applied to the object
     srfp_channel = 0xFFFFFFFE, // Use surface parameterization.
-    emap_channel = 0xFFFFFFFF  // Environment map the geometric object.
+    emap_channel = (long)0xFFFFFFFF  // Environment map the geometric object.
   };
 
   // If the m_mapping_channel_id value is one of the built-in 
@@ -169,7 +169,9 @@ public:
     // OBSOLETE - set m_mapping_channel_id = ON_MappingChannel::emap_mapping
     emap_texture = 86, // spherical environment mapping.
 
-    force_32bit_texture_type = 0xFFFFFFFF
+    // andrew: case value evaluates to 4294967295, which cannot be narrowed to type 'int' [-Wc++11-narrowing]
+    //force_32bit_texture_type = 0xFFFFFFFF
+    force_32bit_texture_type = (long)0xFFFFFFFF
   };
 
   TYPE m_type;
@@ -185,7 +187,7 @@ public:
                            // To "add" a texture, set m_blend_amount = +1
                            // To "subtract" a texture, set m_blend_amount = -1
 
-    force_32bit_texture_mode = 0xFFFFFFFF
+    force_32bit_texture_mode = (long)0xFFFFFFFF
   };
 
   MODE m_mode;
@@ -195,7 +197,7 @@ public:
     nearest_filter = 0, // nearest texture pixel is used
     linear_filter  = 1, // weighted average of corresponding texture pixels
 
-    force_32bit_texture_filter = 0xFFFFFFFF
+    force_32bit_texture_filter = (long)0xFFFFFFFF
   };
   
   // The value of m_minfilter determines how the color
@@ -213,7 +215,7 @@ public:
     repeat_wrap      = 0,
     clamp_wrap       = 1,
 
-    force_32bit_texture_wrap = 0xFFFFFFFF
+    force_32bit_texture_wrap = (long)0xFFFFFFFF
   };
 
   WRAP m_wrapu;
